Added Map Generator
Added the map generator back from an old revision. It works, but doesnt
integrate nicely with my changes. I've listed some modifications that
need to be made in TODO to make it play nicely with the rest of the
engine.

Also I'm now realizing how clunky and horribly my movement, and by
extension, animation system is. This needs to be overhauled.